requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mysqlclient==2.2.4
pendulum==3.0.0
PyJWT==2.7.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
time-machine==2.14.1
twilio==9.0.5
typing_extensions==4.6.2
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
